backend/pyproject.toml
[project]
name = "question-ranker-backend"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0",
    "djangorestframework>=3.15.2",
    "djangorestframework-simplejwt>=5.3.1",
    "django-cors-headers>=4.8.0",
    "numpy>=2.4.0",
    "openai>=2.14.0",
    "python-dotenv>=1.2.1",
]
